requirements.txt
streamlit
pandas
plotly.express
seaborn
numpy
matplotlib.pyplot